ruff and moved one test
