+++ tests/test_alias.py
@@ -143,6 +143,24 @@ class TestAlias:
         assert config.data.attr_field == 100
         assert config.data.str_field == "Test"
 
+    def test_extra_data_nested_aliased(self):
+        @dataclass
+        class ConfigAliasedParent:
+            import_: Config42 = field(
+                default_factory=lambda: Config42(), metadata={"alias": "import"}
+            )
+
+        config = ConfigExtra(ConfigAliasedParent())
+        assert config.data.import_.int_field == 42
+        assert config.data["import"].int_field == 42
+
+        config.data.import_.new_field = "New Value"
+        assert config.data.import_.new_field == "New Value"
+        assert config.data._extra_data["import"]["new_field"] == "New Value"
+        assert config._extra_data["import"]["new_field"] == "New Value"
+        assert config.data["import"].new_field == "New Value"
+        assert config.data["import"]["new_field"] == "New Value"
+
 
 class TestToDictAlias:
     """Test that aliases are resolved properly when using the to_dict function."""
