+++ tests/test_config_extra_fields.py
@@ -1,4 +1,3 @@
-from copy import deepcopy
 from dataclasses import dataclass, field
 import pytest
 from eyconf.config import ConfigExtra
@@ -48,24 +47,6 @@ class TestDataProperties:
         assert isinstance(conf42.data._extra_data, dict)
         assert conf42.data._extra_data["new_field"] == "New Value"
         assert conf42.data.new_field == "New Value"
-
-    def test_extra_data_nested_aliased(self):
-        @dataclass
-        class ConfigAliasedParent:
-            import_: Config42 = field(
-                default_factory=lambda: Config42(), metadata={"alias": "import"}
-            )
-
-        config = ConfigExtra(ConfigAliasedParent())
-        assert config.data.import_.int_field == 42
-        assert config.data["import"].int_field == 42
-
-        config.data.import_.new_field = "New Value"
-        assert config.data.import_.new_field == "New Value"
-        assert config.data._extra_data["import"]["new_field"] == "New Value"
-        assert config._extra_data["import"]["new_field"] == "New Value"
-        assert config.data["import"].new_field == "New Value"
-        assert config.data["import"]["new_field"] == "New Value"
 
 
 class TestUpdate:
